+++ src/app/(main)/my-voyage/[teamId]/ideation/adapters/ideationSA.ts
@@ -4,7 +4,8 @@ import { revalidateTag } from "next/cache";
 import { IdeationClientAdapter } from "./ideationClientAdapter";
 import { getAccessToken } from "@/utils/getCookie";
 import { DELETE, PATCH, POST } from "@/utils/requests";
-import { type AsyncActionResponse, handleAsync } from "@/utils/handleAsync";
+import { handleAsync } from "@/utils/handleAsync";
+import { type AsyncActionResponse } from "@/modules/shared/types";
 import { CacheTag } from "@/utils/cacheTag";
 import { type AddIdeationResponseDto } from "@/modules/ideation/application/dtos/response.dto";
 import { type AddIdeationUsecaseDto } from "@/modules/ideation/application/dtos/addIdeationUsecaseDto";
