+++ src/modules/ideation/ports/primary/ideationClientPort.ts
@@ -1,5 +1,5 @@
 import { type AddIdeationUsecaseDto } from "@/modules/ideation/application/dtos/addIdeationUsecaseDto";
-import { type AsyncActionResponse } from "@/utils/handleAsync";
+import { type AsyncActionResponse } from "@/modules/shared/types";
 import { type AddIdeationResponseDto } from "@/modules/ideation/application/dtos/response.dto";
 
 export interface IdeationClientPort {
