+++ src/types/types.ts
@@ -1,5 +1,1 @@
 // only put global types here
-
-export interface AppError {
-  message: string;
-}
